KYAN-154 Possibility to set a background image for a card component
