+++ applications/common/backend/src/main/resources/libs/kyanite/common/components/common/backgroundimage/.content.json
@@ -1,0 +1,46 @@
+{
+  "sling:resourceType": "wcm/dialogs/components/tab",
+  "label": "Background image",
+  "desktopBackgroundImage": {
+    "sling:resourceType": "wcm/dialogs/components/include",
+    "include": {
+      "sling:resourceSuperType": "kyanite/common/components/common/imagesource",
+      "name": "Desktop background image",
+      "assetReference": {
+        "mimeTypes": ["image/*"],
+        "name": "desktopBackgroundImage/assetReference"
+      },
+      "src": {
+        "name": "desktopBackgroundImage/assetReference"
+      }
+    }
+  },
+  "tabletBackgroundImage": {
+    "sling:resourceType": "wcm/dialogs/components/include",
+    "include": {
+      "sling:resourceSuperType": "kyanite/common/components/common/imagesource",
+      "name": "Tablet background image",
+      "assetReference": {
+        "mimeTypes": ["image/*"],
+        "name": "tabletBackgroundImage/assetReference"
+      },
+      "src": {
+        "name": "tabletBackgroundImage/assetReference"
+      }
+    }
+  },
+  "mobileBackgroundImage": {
+    "sling:resourceType": "wcm/dialogs/components/include",
+    "include": {
+      "sling:resourceSuperType": "kyanite/common/components/common/imagesource",
+      "name": "Mobile background image",
+      "assetReference": {
+        "mimeTypes": ["image/*"],
+        "name": "mobileBackgroundImage/assetReference"
+      },
+      "src": {
+        "name": "mobileBackgroundImage/assetReference"
+      }
+    }
+  }
+}
