+++ applications/common/backend/src/main/resources/libs/kyanite/common/components/section/dialog/.content.json
@@ -17,50 +17,8 @@
 			}
 		},
 		"backgroundImage": {
-			"sling:resourceType": "wcm/dialogs/components/tab",
-			"label": "Background image",
-			"desktopBackgroundImage": {
-				"sling:resourceType": "wcm/dialogs/components/include",
-				"include": {
-					"sling:resourceSuperType": "kyanite/common/components/common/imagesource",
-					"name": "Desktop background image",
-					"assetReference": {
-						"mimeTypes": ["image/*"],
-						"name": "desktopBackgroundImage/assetReference"
-					},
-					"src": {
-						"name": "desktopBackgroundImage/assetReference"
-					}
-				}
-			},
-			"tabletBackgroundImage": {
-				"sling:resourceType": "wcm/dialogs/components/include",
-				"include": {
-					"sling:resourceSuperType": "kyanite/common/components/common/imagesource",
-					"name": "Tablet background image",
-					"assetReference": {
-						"mimeTypes": ["image/*"],
-						"name": "tabletBackgroundImage/assetReference"
-					},
-					"src": {
-						"name": "tabletBackgroundImage/assetReference"
-					}
-				}
-			},
-			"mobileBackgroundImage": {
-				"sling:resourceType": "wcm/dialogs/components/include",
-				"include": {
-					"sling:resourceSuperType": "kyanite/common/components/common/imagesource",
-					"name": "Mobile background image",
-					"assetReference": {
-						"mimeTypes": ["image/*"],
-						"name": "mobileBackgroundImage/assetReference"
-					},
-					"src": {
-						"name": "mobileBackgroundImage/assetReference"
-					}
-				}
-			}
+			"sling:resourceType": "wcm/dialogs/components/include",
+			"path": "/libs/kyanite/common/components/common/backgroundimage"
 		}
   }
 }
