+++ applications/common/frontend/src/components/Card/Card.scss
@@ -16,6 +16,7 @@
  * -->
  */
 
+@use 'src/atomic-design-system/01-atom/a.bg-image';
 @use 'src/atomic-design-system/02-molecule/m.card' as card;
 @use 'src/atomic-design-system/02-molecule/m.card-full-height' as
   card-full-height;
@@ -25,5 +26,9 @@
 
   &--full-height {
     @include card-full-height.m-card-full-height;
+  }
+
+  &.with-background-image {
+    @extend .a-bg-image
   }
 }
